Split progress::manager into a separate file
